indentation

+++ Headers/Arch/InterruptDescriptorTable.hpp
@@ -9,7 +9,7 @@
 #include "InterruptServiceRoutine/InterruptServiceRoutineEntries.hpp"
 #include "InterruptServiceRoutine/InterruptServiceRoutineDefenitions.hpp"
 
-extern "C" ProcessorRegisterSet * isr_function_handler(ProcessorRegisterSet * registers, InterruptServiceRoutineEntries * isr_entries);
+extern "C" ProcessorRegisterSet * isr_function_handler(ProcessorRegisterSet *registers, InterruptServiceRoutineEntries *isr_entries);
 
 struct __attribute__((packed)) IdtEntry
 {
@@ -38,7 +38,7 @@ public:
 
 public:
 	Span<IdtEntry> getEntries();
-	IdtDescriptor* getIdtDescriptor();
+	IdtDescriptor * getIdtDescriptor();
 
 private:
 	void idt_set_entry(uint8_t entry_number, IsrFunction_t* funcall);
@@ -46,7 +46,7 @@ private:
 
 private:
 	Span<IdtEntry> _entries;
-	constexpr static uint16_t IDT_SIZE = 256;
+	constexpr static size_t IDT_SIZE = 256;
 	constexpr static uint16_t NO_INTERRUPT_HANDLER = 0;
 	alignas(uint16_t) IdtDescriptor* _idt_descriptor;
 };
